Move rust_manifest crate into rust_install
