+++ rust-install/Cargo.toml
@@ -20,7 +20,6 @@ itertools = "0.4.1"
 ole32-sys = "0.2.0"
 tar = "0.4.0"
 flate2 = "0.2.9"
-rust-manifest = { version = "0.0.5", path = "../rust-manifest" }
 tempdir = "0.3.4"
 walkdir = "0.1.5"
 toml = "0.1.27"
