+++ rust-install/src/dist.rs
@@ -3,8 +3,8 @@ use temp;
 use errors::*;
 use utils;
 use InstallPrefix;
-use rust_manifest::Component;
-use rust_manifest::Manifest as ManifestV2;
+use manifest::Component;
+use manifest::Manifest as ManifestV2;
 use manifestation::{Manifestation, UpdateStatus, Changes};
 use hyper;
