--- rust-manifest/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-
-name = "rust-manifest"
-version = "0.0.5"
-authors = [ "Diggory Blake <diggsey@googlemail.com>" ]
-description = "Serialization of rust manifest files"
-
-documentation = "http://diggsey.github.io/multirust-rs/rust_manifest/index.html"
-homepage = "https://github.com/Diggsey/multirust-rs"
-repository = "https://github.com/Diggsey/multirust-rs"
-
-license = "MIT OR Apache-2.0"
-
-[dependencies]
-toml = "0.1.23"
-
-[lib]
-name = "rust_manifest"
