--- rust-manifest/src/lib.rs
@@ -1,10 +1,0 @@
-extern crate toml;
-
-pub use errors::*;
-pub use manifest::*;
-pub use config::*;
-
-mod errors;
-mod utils;
-mod manifest;
-mod config;
